Use BigDecimal to calculate foreign currency history. Refactor create balance entry process.
